requirements.txt
python-dotenv==0.19.0
selenium==3.141.0
SQLAlchemy==1.4.23
psycopg2==2.9.1
pandas==1.3.2
